PRSD-659: Refactors landlord search integration tests to prepare for property ones
