+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/SearchLandlordRegisterPage.kt
@@ -2,38 +2,19 @@ package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages
 
 import com.microsoft.playwright.Page
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getChildComponent
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.BaseComponent.Companion.getComponent
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.FilterPanel
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Pagination
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.SearchBar
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Table
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.SearchRegisterBasePage
 
 class SearchLandlordRegisterPage(
     page: Page,
-) : BasePage(page, "/search/landlord") {
-    val searchBar = SearchBar(page)
-
-    fun getFilterPanel() = FilterPanel(page)
-
-    fun getResultTable() = Table(page)
-
+) : SearchRegisterBasePage(page, "/search/landlord") {
     fun getLandlordLink(rowIndex: Int) = getChildComponent(getResultTable().getCell(rowIndex, LANDLORD_COL_INDEX), "a")
 
-    fun getPaginationComponent() = Pagination(page)
-
-    fun getErrorMessageText() = getErrorMessage().innerText()
-
     fun getPropertySearchLink() = getChildComponent(getErrorMessage(), "a")
-
-    fun getErrorMessage() = getComponent(page, "#no-results")
 
     companion object {
         const val LANDLORD_COL_INDEX: Int = 0
         const val ADDRESS_COL_INDEX: Int = 1
         const val CONTACT_INFO_COL_INDEX: Int = 2
         const val LISTED_PROPERTY_COL_INDEX: Int = 3
-
-        const val LA_FILTER_INDEX = 0
     }
 }
